+++ errors.h
@@ -108,7 +108,6 @@ typedef enum {
 
 typedef struct {
     status_code_t id;
-    const char *xname;
     const char *description;
 } status_detail_t;